refactor: move download helpers
