+++ app/src/main/java/org/ole/planet/myplanet/utilities/DownloadUtils.kt
@@ -1,15 +1,24 @@
 package org.ole.planet.myplanet.utilities
 
+import android.app.ActivityManager
+import android.app.AlarmManager
 import android.app.Notification
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
 import android.os.Build
+import androidx.annotation.RequiresApi
 import androidx.core.app.NotificationCompat
+import androidx.core.content.edit
+import androidx.work.OneTimeWorkRequestBuilder
+import androidx.work.WorkManager
+import androidx.work.workDataOf
 import java.util.regex.Pattern
 import kotlin.text.isNotEmpty
 import org.ole.planet.myplanet.MainApplication
 import org.ole.planet.myplanet.R
+import org.ole.planet.myplanet.datamanager.DownloadWorker
+import org.ole.planet.myplanet.datamanager.MyDownloadService
 import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.utilities.FileUtils
 
@@ -132,6 +141,87 @@ object DownloadUtils {
             urls.add(Utilities.getUrl(dbMyLibrary[selectedItems[i]]))
         }
         return urls
+    }
+
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun openDownloadService(context: Context?, urls: ArrayList<String>, fromSync: Boolean) {
+        context?.let { ctx ->
+            val preferences = ctx.getSharedPreferences(MyDownloadService.PREFS_NAME, Context.MODE_PRIVATE)
+            preferences.edit {
+                putStringSet("url_list_key", urls.toSet())
+            }
+            startDownloadServiceSafely(ctx, "url_list_key", fromSync)
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.S)
+    private fun startDownloadServiceSafely(context: Context, urlsKey: String, fromSync: Boolean) {
+        if (canStartForegroundService(context)) {
+            try {
+                MyDownloadService.startService(context, urlsKey, fromSync)
+            } catch (e: Exception) {
+                e.printStackTrace()
+                handleForegroundServiceNotAllowed(context, urlsKey, fromSync)
+            }
+        } else {
+            handleForegroundServiceNotAllowed(context, urlsKey, fromSync)
+        }
+    }
+
+    private fun handleForegroundServiceNotAllowed(context: Context, urlsKey: String, fromSync: Boolean) {
+        if (!fromSync) {
+            Utilities.toast(context, context.getString(R.string.download_in_background))
+        }
+        startDownloadWork(context, urlsKey, fromSync)
+    }
+
+    private fun startDownloadWork(context: Context, urlsKey: String, fromSync: Boolean) {
+        val workRequest = OneTimeWorkRequestBuilder<DownloadWorker>()
+            .setInputData(
+                workDataOf(
+                    "urls_key" to urlsKey,
+                    "fromSync" to fromSync
+                )
+            )
+            .addTag("download_work")
+            .build()
+
+        WorkManager.getInstance(context).enqueue(workRequest)
+    }
+
+    fun canStartForegroundService(context: Context): Boolean {
+        return when {
+            Build.VERSION.SDK_INT < Build.VERSION_CODES.O -> true
+            Build.VERSION.SDK_INT < Build.VERSION_CODES.S -> {
+                isAppInForeground(context)
+            }
+            else -> {
+                isAppInForeground(context) || hasSpecialForegroundPermissions(context)
+            }
+        }
+    }
+
+    private fun isAppInForeground(context: Context): Boolean {
+        val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
+        val appProcesses = activityManager.runningAppProcesses ?: return false
+
+        val packageName = context.packageName
+        return appProcesses.any { processInfo ->
+            processInfo.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND &&
+                processInfo.processName == packageName
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.S)
+    private fun hasSpecialForegroundPermissions(context: Context): Boolean {
+        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+        return when {
+            Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+                alarmManager.canScheduleExactAlarms()
+            }
+            else -> false
+        }
     }
 
     fun extractLinks(text: String?): ArrayList<String> {
