+++ app/src/main/java/org/ole/planet/myplanet/utilities/Utilities.kt
@@ -1,11 +1,8 @@
 package org.ole.planet.myplanet.utilities
 
-import android.app.ActivityManager
-import android.app.AlarmManager
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
-import android.os.Build
 import android.os.Handler
 import android.os.Looper
 import android.text.format.DateUtils
@@ -15,21 +12,14 @@ import android.util.Patterns
 import android.webkit.MimeTypeMap
 import android.widget.ImageView
 import android.widget.Toast
-import androidx.annotation.RequiresApi
-import androidx.core.content.edit
 import androidx.core.graphics.toColorInt
 import androidx.core.net.toUri
-import androidx.work.OneTimeWorkRequestBuilder
-import androidx.work.WorkManager
-import androidx.work.workDataOf
 import com.bumptech.glide.Glide
 import fisk.chipcloud.ChipCloudConfig
 import java.lang.ref.WeakReference
 import java.math.BigInteger
 import org.ole.planet.myplanet.MainApplication.Companion.context
 import org.ole.planet.myplanet.R
-import org.ole.planet.myplanet.datamanager.DownloadWorker
-import org.ole.planet.myplanet.datamanager.MyDownloadService
 import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.utilities.Constants.PREFS_NAME
 import org.ole.planet.myplanet.utilities.UrlUtils
@@ -64,86 +54,6 @@ object Utilities {
 
     fun getUserImageUrl(userId: String?, imageName: String): String {
         return "${getUrl()}/_users/$userId/$imageName"
-    }
-
-    @RequiresApi(Build.VERSION_CODES.S)
-    fun openDownloadService(context: Context?, urls: ArrayList<String>, fromSync: Boolean) {
-        context?.let { ctx ->
-            val preferences = ctx.getSharedPreferences(MyDownloadService.PREFS_NAME, Context.MODE_PRIVATE)
-            preferences.edit {
-                putStringSet("url_list_key", urls.toSet())
-            }
-            startDownloadServiceSafely(ctx, "url_list_key", fromSync)
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.S)
-    private fun startDownloadServiceSafely(context: Context, urlsKey: String, fromSync: Boolean) {
-        if (canStartForegroundService(context)) {
-            try {
-                MyDownloadService.startService(context, urlsKey, fromSync)
-            } catch (e: Exception) {
-                e.printStackTrace()
-                handleForegroundServiceNotAllowed(context, urlsKey, fromSync)
-            }
-        } else {
-            handleForegroundServiceNotAllowed(context, urlsKey, fromSync)
-        }
-    }
-
-    private fun handleForegroundServiceNotAllowed(context: Context, urlsKey: String, fromSync: Boolean) {
-        if (!fromSync) {
-            toast(context, context.getString(R.string.download_in_background))
-        }
-        startDownloadWork(context, urlsKey, fromSync)
-    }
-
-    private fun startDownloadWork(context: Context, urlsKey: String, fromSync: Boolean) {
-        val workRequest = OneTimeWorkRequestBuilder<DownloadWorker>()
-            .setInputData(workDataOf(
-                "urls_key" to urlsKey,
-                "fromSync" to fromSync
-            ))
-            .addTag("download_work")
-            .build()
-
-        WorkManager.getInstance(context).enqueue(workRequest)
-    }
-
-    fun canStartForegroundService(context: Context): Boolean {
-        return when {
-            Build.VERSION.SDK_INT < Build.VERSION_CODES.O -> true
-            Build.VERSION.SDK_INT < Build.VERSION_CODES.S -> {
-                // Android 8-11: Check if app is in foreground or has special permissions
-                isAppInForeground(context)
-            }
-            else -> {
-                // Android 12+: More restrictions
-                isAppInForeground(context) || hasSpecialForegroundPermissions(context)
-            }
-        }
-    }
-
-    private fun isAppInForeground(context: Context): Boolean {
-        val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
-        val appProcesses = activityManager.runningAppProcesses ?: return false
-
-        val packageName = context.packageName
-        return appProcesses.any { processInfo ->
-            processInfo.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND && processInfo.processName == packageName
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.S)
-    private fun hasSpecialForegroundPermissions(context: Context): Boolean {
-        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-
-        return when {
-            Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
-                alarmManager.canScheduleExactAlarms()
-            }
-            else -> false
-        }
     }
 
     @JvmStatic
